Pattern Match: split large file into smaller parts
